+++ bsp/cvitek/cv18xx_aarch64/board/board.c
@@ -89,15 +89,3 @@ void rt_hw_board_init(void)
     rt_hw_common_setup();
 }
 #endif /* RT_USING_OFW */
-
-static rt_ubase_t pinmux_base = RT_NULL;
-
-rt_ubase_t pinmux_base_ioremap(void)
-{
-    if (pinmux_base == RT_NULL)
-    {
-        pinmux_base = (rt_size_t)rt_ioremap((void*)0x03001000, 0x1000);
-    }
-
-    return pinmux_base;
-}
