+++ bsp/cvitek/drivers/drv_pinmux.c
@@ -1,0 +1,14 @@
+#include <rtthread.h>
+#include <rtioremap.h>
+
+static rt_ubase_t pinmux_base = RT_NULL;
+
+rt_ubase_t pinmux_base_ioremap(void)
+{
+    if (pinmux_base == RT_NULL)
+    {
+        pinmux_base = (rt_size_t)rt_ioremap((void*)0x03001000, 0x1000);
+    }
+
+    return pinmux_base;
+}
